изменен тип данных идшника

+++ TP-CP-5-Semester/RequestBodies/OrderRequest.cs
@@ -3,6 +3,6 @@ namespace TP_CP_5_Semester.RequestBodies;
 public class OrderRequest
 {
     public string Email { get; set; }
-    public long TourId { get; set; }
+    public int TourId { get; set; }
     public int Amount { get; set; }
 }
